Late-Fall Run Chinook Model

Late-Fall Run Chinook life cycle model used for CVPIA's Structured Decision Making Process

late_fall_run_model(
  scenario = NULL,
  mode = c("seed", "simulate", "calibrate"),
  seeds = NULL,
  ..params = lateFallRunDSM::params,
  stochastic = FALSE
)

Source

IP-117068

Arguments

scenario

Model inputs, can be modified to test management actions

mode

The mode to run model in. Can be "seed", "simulate" or "calibrate".

seeds

The default value is NULL runs the model in seeding mode, returning a 31 by 25 matrix with the first four years of seeded adults. This returned value can be fed into the model again as the value for the seeds argument

..params

Parameters for model and submodels. More details at params

Examples

late_fall_run_seeds <- lateFallRunDSM::late_fall_run_model(mode = "seed")
lateFallRunDSM::late_fall_run_model(scenario = DSMscenario::scenarios$ONE,
                           mode = "simulate",
                           seeds = late_fall_run_seeds)
#> $spawners
#>                               1    2    3    4    5    6    7    8    9   10
#> Upper Sacramento River     4821 5458 3852 6162 4099 5772 6855 2428 2251 2061
#> Antelope Creek                4    4    4    7    4    5    6    2    2    2
#> Battle Creek                 64   66   69   73   64   65   66   64   64   68
#> Bear Creek                    4    4    4    7    4    5    6    2    2    2
#> Big Chico Creek               4    4    4    7    4    5    6    2    2    2
#> Butte Creek                   4    4    4    7    4    5    6    2    2    2
#> Clear Creek                 252  296  324  638  437  564  656  224  224  226
#> Cottonwood Creek              4    4    4    7    4    5    6    2    2    2
#> Cow Creek                     4    4    4    7    4    5    6    2    2    2
#> Deer Creek                    4    4    4    7    4    5    6    2    2    2
#> Elder Creek                   4    4    4    7    4    5    6    2    2    2
#> Mill Creek                    4    4    4    7    4    5    6    2    2    2
#> Paynes Creek                  4    4    4    7    4    5    6    2    2    2
#> Stony Creek                   4    4    4    7    4    5    6    2    2    2
#> Thomes Creek                  4    4    4    7    4    5    6    2    2    2
#> Upper-mid Sacramento River    0    0    0    0    0    0    0    0    0    0
#> Sutter Bypass                 0    0    0    0    0    0    0    0    0    0
#> Bear River                    4    4    4    7    4    5    6    2    2    2
#> Feather River                 4    4    4    7    4    5    6    2    2    2
#> Yuba River                    4    4    4    7    4    5    6    2    2    2
#> Lower-mid Sacramento River    0    0    0    0    0    0    0    0    0    0
#> Yolo Bypass                   0    0    0    0    0    0    0    0    0    0
#> American River                4    4    4    7    4    5    6    2    2    2
#> Lower Sacramento River        0    0    0    0    0    0    0    0    0    0
#> Calaveras River               4    4    4    7    4    5    6    2    2    2
#> Cosumnes River                4    4    4    7    4    5    6    2    2    2
#> Mokelumne River               4    4    4    7    4    5    6    2    2    2
#> Merced River                  4    4    4    7    4    5    6    2    2    2
#> Stanislaus River              4    4    4    7    4    5    6    2    2    2
#> Tuolumne River                4    4    4    7    4    5    6    2    2    2
#> San Joaquin River             0    0    0    0    0    0    0    0    0    0
#>                              11   12   13  14  15  16  17  18  19   20
#> Upper Sacramento River     2121 1380 1067 412 393 518 840 972 994 1152
#> Antelope Creek                2    1    0   0   0   0   0   0   0    1
#> Battle Creek                 70   69   68  58  58  68  76  74  66   61
#> Bear Creek                    2    1    0   0   0   0   0   0   0    1
#> Big Chico Creek               2    1    0   0   0   0   0   0   0    1
#> Butte Creek                   2    1    0   0   0   0   0   0   0    1
#> Clear Creek                 294  205  146  26  34  43  90 116 136  213
#> Cottonwood Creek              2    1    0   0   0   0   0   0   0    1
#> Cow Creek                     2    1    0   0   0   0   0   0   0    1
#> Deer Creek                    2    1    0   0   0   0   0   0   0    1
#> Elder Creek                   2    1    0   0   0   0   0   0   0    1
#> Mill Creek                    2    1    0   0   0   0   0   0   0    1
#> Paynes Creek                  2    1    0   0   0   0   0   0   0    1
#> Stony Creek                   2    1    0   0   0   0   0   0   0    1
#> Thomes Creek                  2    1    0   0   0   0   0   0   0    1
#> Upper-mid Sacramento River    0    0    0   0   0   0   0   0   0    0
#> Sutter Bypass                 0    0    0   0   0   0   0   0   0    0
#> Bear River                    2    1    0   0   0   0   0   0   0    1
#> Feather River                 2    1    0   0   0   0   0   0   0    1
#> Yuba River                    2    1    0   0   0   0   0   0   0    1
#> Lower-mid Sacramento River    0    0    0   0   0   0   0   0   0    0
#> Yolo Bypass                   0    0    0   0   0   0   0   0   0    0
#> American River                2    1    0   0   0   0   0   0   0    1
#> Lower Sacramento River        0    0    0   0   0   0   0   0   0    0
#> Calaveras River               2    1    0   0   0   0   0   0   0    1
#> Cosumnes River                2    1    0   0   0   0   0   0   0    1
#> Mokelumne River               2    1    0   0   0   0   0   0   0    1
#> Merced River                  2    1    0   0   0   0   0   0   0    1
#> Stanislaus River              2    1    0   0   0   0   0   0   0    1
#> Tuolumne River                2    1    0   0   0   0   0   0   0    1
#> San Joaquin River             0    0    0   0   0   0   0   0   0    0
#> 
#> $juvenile_biomass
#>                                     1           2         3           4
#> Upper Sacramento River     291471.833 288127.9000 70667.600 788193.1333
#> Antelope Creek                  0.000      0.0000     0.000      0.0000
#> Battle Creek                 1308.333    450.0667   125.600    209.3333
#> Bear Creek                      0.000      0.0000     0.000      0.0000
#> Big Chico Creek                 0.000      0.0000     0.000      0.0000
#> Butte Creek                     0.000      0.0000     0.000      0.0000
#> Clear Creek                 35301.200  39739.7667  8625.467  74583.5333
#> Cottonwood Creek                0.000      0.0000     0.000      0.0000
#> Cow Creek                       0.000      0.0000     0.000      0.0000
#> Deer Creek                      0.000      0.0000     0.000      0.0000
#> Elder Creek                     0.000      0.0000     0.000      0.0000
#> Mill Creek                      0.000      0.0000     0.000      0.0000
#> Paynes Creek                    0.000      0.0000     0.000      0.0000
#> Stony Creek                     0.000      0.0000     0.000      0.0000
#> Thomes Creek                    0.000      0.0000     0.000      0.0000
#> Upper-mid Sacramento River      0.000      0.0000     0.000      0.0000
#> Sutter Bypass                   0.000      0.0000     0.000      0.0000
#> Bear River                      0.000      0.0000     0.000      0.0000
#> Feather River                   0.000      0.0000     0.000      0.0000
#> Yuba River                      0.000      0.0000     0.000      0.0000
#> Lower-mid Sacramento River      0.000      0.0000     0.000      0.0000
#> Yolo Bypass                     0.000      0.0000     0.000      0.0000
#> American River                  0.000      0.0000     0.000      0.0000
#> Lower Sacramento River          0.000      0.0000     0.000      0.0000
#> Calaveras River                 0.000      0.0000     0.000      0.0000
#> Cosumnes River                  0.000      0.0000     0.000      0.0000
#> Mokelumne River                 0.000      0.0000     0.000      0.0000
#> Merced River                    0.000      0.0000     0.000      0.0000
#> Stanislaus River                0.000      0.0000     0.000      0.0000
#> Tuolumne River                  0.000      0.0000     0.000      0.0000
#> San Joaquin River               0.000      0.0000     0.000      0.0000
#>                                   5        6           7          8          9
#> Upper Sacramento River     87620.57 218680.0 52548.23333 253513.800 71636.4333
#> Antelope Creek                 0.00      0.0     0.00000      0.000     0.0000
#> Battle Creek                 188.40    722.2    73.26667   1433.933   826.8667
#> Bear Creek                     0.00      0.0     0.00000      0.000     0.0000
#> Big Chico Creek                0.00      0.0     0.00000      0.000     0.0000
#> Butte Creek                    0.00      0.0     0.00000      0.000     0.0000
#> Clear Creek                 7957.40  30528.8  5388.03333  39551.867 13352.4667
#> Cottonwood Creek               0.00      0.0     0.00000      0.000     0.0000
#> Cow Creek                      0.00      0.0     0.00000      0.000     0.0000
#> Deer Creek                     0.00      0.0     0.00000      0.000     0.0000
#> Elder Creek                    0.00      0.0     0.00000      0.000     0.0000
#> Mill Creek                     0.00      0.0     0.00000      0.000     0.0000
#> Paynes Creek                   0.00      0.0     0.00000      0.000     0.0000
#> Stony Creek                    0.00      0.0     0.00000      0.000     0.0000
#> Thomes Creek                   0.00      0.0     0.00000      0.000     0.0000
#> Upper-mid Sacramento River     0.00      0.0     0.00000      0.000     0.0000
#> Sutter Bypass                  0.00      0.0     0.00000      0.000     0.0000
#> Bear River                     0.00      0.0     0.00000      0.000     0.0000
#> Feather River                  0.00      0.0     0.00000      0.000     0.0000
#> Yuba River                     0.00      0.0     0.00000      0.000     0.0000
#> Lower-mid Sacramento River     0.00      0.0     0.00000      0.000     0.0000
#> Yolo Bypass                    0.00      0.0     0.00000      0.000     0.0000
#> American River                 0.00      0.0     0.00000      0.000     0.0000
#> Lower Sacramento River         0.00      0.0     0.00000      0.000     0.0000
#> Calaveras River                0.00      0.0     0.00000      0.000     0.0000
#> Cosumnes River                 0.00      0.0     0.00000      0.000     0.0000
#> Mokelumne River                0.00      0.0     0.00000      0.000     0.0000
#> Merced River                   0.00      0.0     0.00000      0.000     0.0000
#> Stanislaus River               0.00      0.0     0.00000      0.000     0.0000
#> Tuolumne River                 0.00      0.0     0.00000      0.000     0.0000
#> San Joaquin River              0.00      0.0     0.00000      0.000     0.0000
#>                                    10         11       12         13        14
#> Upper Sacramento River     125295.967 2352.06667 6093.300 1755.63333 41982.833
#> Antelope Creek                  0.000    0.00000    0.000    0.00000     0.000
#> Battle Creek                 1465.333   20.93333    0.000   20.93333  1925.867
#> Bear Creek                      0.000    0.00000    0.000    0.00000     0.000
#> Big Chico Creek                 0.000    0.00000    0.000    0.00000     0.000
#> Butte Creek                     0.000    0.00000    0.000    0.00000     0.000
#> Clear Creek                 24377.167  494.20000 2257.833 1417.43333  6819.400
#> Cottonwood Creek                0.000    0.00000    0.000    0.00000     0.000
#> Cow Creek                       0.000    0.00000    0.000    0.00000     0.000
#> Deer Creek                      0.000    0.00000    0.000    0.00000     0.000
#> Elder Creek                     0.000    0.00000    0.000    0.00000     0.000
#> Mill Creek                      0.000    0.00000    0.000    0.00000     0.000
#> Paynes Creek                    0.000    0.00000    0.000    0.00000     0.000
#> Stony Creek                     0.000    0.00000    0.000    0.00000     0.000
#> Thomes Creek                    0.000    0.00000    0.000    0.00000     0.000
#> Upper-mid Sacramento River      0.000    0.00000    0.000    0.00000     0.000
#> Sutter Bypass                   0.000    0.00000    0.000    0.00000     0.000
#> Bear River                      0.000    0.00000    0.000    0.00000     0.000
#> Feather River                   0.000    0.00000    0.000    0.00000     0.000
#> Yuba River                      0.000    0.00000    0.000    0.00000     0.000
#> Lower-mid Sacramento River      0.000    0.00000    0.000    0.00000     0.000
#> Yolo Bypass                     0.000    0.00000    0.000    0.00000     0.000
#> American River                  0.000    0.00000    0.000    0.00000     0.000
#> Lower Sacramento River          0.000    0.00000    0.000    0.00000     0.000
#> Calaveras River                 0.000    0.00000    0.000    0.00000     0.000
#> Cosumnes River                  0.000    0.00000    0.000    0.00000     0.000
#> Mokelumne River                 0.000    0.00000    0.000    0.00000     0.000
#> Merced River                    0.000    0.00000    0.000    0.00000     0.000
#> Stanislaus River                0.000    0.00000    0.000    0.00000     0.000
#> Tuolumne River                  0.000    0.00000    0.000    0.00000     0.000
#> San Joaquin River               0.000    0.00000    0.000    0.00000     0.000
#>                                   15        16         17       18        19
#> Upper Sacramento River     65561.800 40453.267 47212.8667 66816.07 173615.90
#> Antelope Creek                 0.000     0.000     0.0000     0.00      0.00
#> Battle Creek                2302.667   982.500   209.3333   251.20  10979.53
#> Bear Creek                     0.000     0.000     0.0000     0.00      0.00
#> Big Chico Creek                0.000     0.000     0.0000     0.00      0.00
#> Butte Creek                    0.000     0.000     0.0000     0.00      0.00
#> Clear Creek                13123.667  9154.633 12973.0667 17167.47  62657.07
#> Cottonwood Creek               0.000     0.000     0.0000     0.00      0.00
#> Cow Creek                      0.000     0.000     0.0000     0.00      0.00
#> Deer Creek                     0.000     0.000     0.0000     0.00      0.00
#> Elder Creek                    0.000     0.000     0.0000     0.00      0.00
#> Mill Creek                     0.000     0.000     0.0000     0.00      0.00
#> Paynes Creek                   0.000     0.000     0.0000     0.00      0.00
#> Stony Creek                    0.000     0.000     0.0000     0.00      0.00
#> Thomes Creek                   0.000     0.000     0.0000     0.00      0.00
#> Upper-mid Sacramento River     0.000     0.000     0.0000     0.00      0.00
#> Sutter Bypass                  0.000     0.000     0.0000     0.00      0.00
#> Bear River                     0.000     0.000     0.0000     0.00      0.00
#> Feather River                  0.000     0.000     0.0000     0.00      0.00
#> Yuba River                     0.000     0.000     0.0000     0.00      0.00
#> Lower-mid Sacramento River     0.000     0.000     0.0000     0.00      0.00
#> Yolo Bypass                    0.000     0.000     0.0000     0.00      0.00
#> American River                 0.000     0.000     0.0000     0.00      0.00
#> Lower Sacramento River         0.000     0.000     0.0000     0.00      0.00
#> Calaveras River                0.000     0.000     0.0000     0.00      0.00
#> Cosumnes River                 0.000     0.000     0.0000     0.00      0.00
#> Mokelumne River                0.000     0.000     0.0000     0.00      0.00
#> Merced River                   0.000     0.000     0.0000     0.00      0.00
#> Stanislaus River               0.000     0.000     0.0000     0.00      0.00
#> Tuolumne River                 0.000     0.000     0.0000     0.00      0.00
#> San Joaquin River              0.000     0.000     0.0000     0.00      0.00
#>                                   20
#> Upper Sacramento River     130478.40
#> Antelope Creek                  0.00
#> Battle Creek                 1507.20
#> Bear Creek                      0.00
#> Big Chico Creek                 0.00
#> Butte Creek                     0.00
#> Clear Creek                 46781.27
#> Cottonwood Creek                0.00
#> Cow Creek                       0.00
#> Deer Creek                      0.00
#> Elder Creek                     0.00
#> Mill Creek                      0.00
#> Paynes Creek                    0.00
#> Stony Creek                     0.00
#> Thomes Creek                    0.00
#> Upper-mid Sacramento River      0.00
#> Sutter Bypass                   0.00
#> Bear River                      0.00
#> Feather River                   0.00
#> Yuba River                      0.00
#> Lower-mid Sacramento River      0.00
#> Yolo Bypass                     0.00
#> American River                  0.00
#> Lower Sacramento River          0.00
#> Calaveras River                 0.00
#> Cosumnes River                  0.00
#> Mokelumne River                 0.00
#> Merced River                    0.00
#> Stanislaus River                0.00
#> Tuolumne River                  0.00
#> San Joaquin River               0.00
#> 
#> $proportion_natural
#>                                    1         2         3         4         5
#> Upper Sacramento River     0.9330015 0.9408208 0.9161475 0.9475820 0.9212003
#> Antelope Creek             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Battle Creek               0.0937500 0.1212121 0.1594203 0.2054795 0.0937500
#> Bear Creek                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Big Chico Creek            1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Butte Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Clear Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Cottonwood Creek           1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Cow Creek                  1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Deer Creek                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Elder Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Mill Creek                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Paynes Creek               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Stony Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Thomes Creek               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sutter Bypass              0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Bear River                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Feather River              1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Yuba River                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Yolo Bypass                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> American River             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Lower Sacramento River     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Calaveras River            1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Cosumnes River             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Mokelumne River            1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Merced River               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Stanislaus River           1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Tuolumne River             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> San Joaquin River          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                                    6         7         8         9        10
#> Upper Sacramento River     0.9440402 0.9528811 0.8669687 0.8565082 0.8432800
#> Antelope Creek             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Battle Creek               0.1076923 0.1212121 0.0937500 0.0937500 0.1470588
#> Bear Creek                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Big Chico Creek            1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Butte Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Clear Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Cottonwood Creek           1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Cow Creek                  1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Deer Creek                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Elder Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Mill Creek                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Paynes Creek               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Stony Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Thomes Creek               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sutter Bypass              0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Bear River                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Feather River              1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Yuba River                 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Yolo Bypass                0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> American River             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Lower Sacramento River     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Calaveras River            1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Cosumnes River             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Mokelumne River            1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Merced River               1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Stanislaus River           1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> Tuolumne River             1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> San Joaquin River          0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>                                   11        12        13        14       15
#> Upper Sacramento River     0.8477133 0.7659420 0.6972821 0.2160194 0.178117
#> Antelope Creek             1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Battle Creek               0.1714286 0.1594203 0.1470588 0.0000000 0.000000
#> Bear Creek                 1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Big Chico Creek            1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Butte Creek                1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Clear Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.000000
#> Cottonwood Creek           1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Cow Creek                  1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Deer Creek                 1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Elder Creek                1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Mill Creek                 1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Paynes Creek               1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Stony Creek                1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Thomes Creek               1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> Sutter Bypass              0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> Bear River                 1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Feather River              1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Yuba River                 1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> Yolo Bypass                0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> American River             1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Lower Sacramento River     0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#> Calaveras River            1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Cosumnes River             1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Mokelumne River            1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Merced River               1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Stanislaus River           1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> Tuolumne River             1.0000000 1.0000000 0.0000000 0.0000000 0.000000
#> San Joaquin River          0.0000000 0.0000000 0.0000000 0.0000000 0.000000
#>                                   16        17        18        19         20
#> Upper Sacramento River     0.3764479 0.6154762 0.6676955 0.6750503 0.71961806
#> Antelope Creek             0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Battle Creek               0.1470588 0.2368421 0.2162162 0.1212121 0.04918033
#> Bear Creek                 0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Big Chico Creek            0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Butte Creek                0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Clear Creek                1.0000000 1.0000000 1.0000000 1.0000000 1.00000000
#> Cottonwood Creek           0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Cow Creek                  0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Deer Creek                 0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Elder Creek                0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Mill Creek                 0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Paynes Creek               0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Stony Creek                0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Thomes Creek               0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Upper-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> Sutter Bypass              0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> Bear River                 0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Feather River              0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Yuba River                 0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Lower-mid Sacramento River 0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> Yolo Bypass                0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> American River             0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Lower Sacramento River     0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> Calaveras River            0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Cosumnes River             0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Mokelumne River            0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Merced River               0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Stanislaus River           0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> Tuolumne River             0.0000000 0.0000000 0.0000000 0.0000000 1.00000000
#> San Joaquin River          0.0000000 0.0000000 0.0000000 0.0000000 0.00000000
#> 
#> $viability_metrics
#>       [,1] [,2] [,3] [,4]
#>  [1,]    0    0    0    0
#>  [2,]    0    1    0    0
#>  [3,]    0    0    0    0
#>  [4,]    0    1    0    0
#>  [5,]    0    0    0    0
#>  [6,]    0    0    0    0
#>  [7,]    0    0    0    0
#>  [8,]    0    0    0    0
#>  [9,]    0    0    0    0
#> [10,]    0    0    0    0
#> [11,]    0    0    0    0
#> [12,]    0    0    0    0
#> [13,]    0    0    0    0
#> [14,]    0    0    0    0
#> [15,]    0    0    0    0
#> [16,]    0    0    0    0
#> [17,]    0    0    0    0
#> [18,]    0    0    0    0
#> [19,]    0    0    0    0
#>